add module formatable to handle all eng/braille formating
