+++ lib/library.rb
@@ -10,7 +10,7 @@ class Library
               "?": ["..", "0.", "00"],
               a:   ["0.", "..", ".."],
               b:   ["0.", "0.", ".."],
-              c:   ["0.", "0.", ".."],
+              c:   ["00", "..", ".."],
               d:   ["00", ".0", ".."],
               e:   ["0.", ".0", ".."],
               f:   ["00", "0.", ".."],
