+++ message.txt
@@ -1,1 +1,1 @@
-hello darkness, my old friend, i've come here to speak with you again! hello darkness, my old friend, i've come here to speak with you again
+hello darkness my old friend i've come here to speak with you again
